extracting test; arm and audioinput component modified
