+++ components/arm/collectors.go
@@ -8,10 +8,8 @@ import (
 	v1 "go.viam.com/api/common/v1"
 	pb "go.viam.com/api/component/arm/v1"
 	"google.golang.org/protobuf/types/known/anypb"
-	"google.golang.org/protobuf/types/known/structpb"
 
 	"go.viam.com/rdk/data"
-	"go.viam.com/rdk/logging"
 	"go.viam.com/rdk/referenceframe"
 	"go.viam.com/rdk/utils"
 )
@@ -105,42 +103,7 @@ func newDoCommandCollector(resource interface{}, params data.CollectorParams) (d
 		return nil, err
 	}
 
-	logger := logging.NewLogger("test")
-	cFunc := data.CaptureFunc(func(ctx context.Context, _ map[string]*anypb.Any) (data.CaptureResult, error) {
-		timeRequested := time.Now()
-		var res data.CaptureResult
-
-		var payload map[string]interface{}
-
-		if payloadAny, exists := params.MethodParams["docommand_input"]; exists && payloadAny != nil {
-			if payloadAny.MessageIs(&structpb.Struct{}) {
-				var s structpb.Struct
-				if err := payloadAny.UnmarshalTo(&s); err != nil {
-					return res, err
-				}
-				payload = s.AsMap()
-			} else {
-				// handle empty payload
-				payload = make(map[string]interface{})
-			}
-		} else {
-			// key does not exist
-			return res, errors.New("missing payload")
-		}
-
-		logger.Info(payload)
-		logger.Info("capturing docommand with payload %#v\n", payload)
-
-		values, err := arm.DoCommand(ctx, payload)
-		if err != nil {
-			if errors.Is(err, data.ErrNoCaptureToStore) {
-				return res, err
-			}
-			return res, data.NewFailedToReadError(params.ComponentName, "DoCommand", err)
-		}
-		ts := data.Timestamps{TimeRequested: timeRequested, TimeReceived: time.Now()}
-		return data.NewTabularCaptureResultDoCommand(ts, values)
-	})
+	cFunc := data.NewDoCommandCaptureFunc(arm, params)
 	return data.NewCollector(cFunc, params)
 }
