+++ components/arm/collectors_test.go
@@ -10,8 +10,6 @@ import (
 	datasyncpb "go.viam.com/api/app/datasync/v1"
 	pb "go.viam.com/api/component/arm/v1"
 	"go.viam.com/test"
-	"google.golang.org/protobuf/types/known/anypb"
-	"google.golang.org/protobuf/types/known/structpb"
 
 	"go.viam.com/rdk/components/arm"
 	"go.viam.com/rdk/data"
@@ -103,91 +101,13 @@ func TestCollectors(t *testing.T) {
 }
 
 func TestDoCommandCollector(t *testing.T) {
-	tests := []struct {
-		name         string
-		collector    data.CollectorConstructor
-		methodParams map[string]*anypb.Any
-		expectError  bool
-	}{
-		{
-			name:      "DoCommand collector should write a list of values",
-			collector: arm.NewDoCommandCollector,
-			methodParams: map[string]*anypb.Any{
-				"docommand_input": func() *anypb.Any {
-					structVal := tu.ToStructPBStruct(t, map[string]any{
-						"command": "random",
-					})
-					anyVal, _ := anypb.New(structVal)
-					return anyVal
-				}(),
-			},
-		},
-		{
-			name:      "DoCommand collector should handle empty struct payload",
-			collector: arm.NewDoCommandCollector,
-			methodParams: map[string]*anypb.Any{
-				"docommand_input": func() *anypb.Any {
-					emptyStruct := &structpb.Struct{
-						Fields: make(map[string]*structpb.Value),
-					}
-					anyVal, _ := anypb.New(emptyStruct)
-					return anyVal
-				}(),
-			},
-		},
-		{
-			name:      "DoCommand collector should handle empty payload",
-			collector: arm.NewDoCommandCollector,
-			methodParams: map[string]*anypb.Any{
-				"docommand_input": {},
-			},
-		},
-		{
-			name:         "DoCommand collector should error on missing payload",
-			collector:    arm.NewDoCommandCollector,
-			methodParams: map[string]*anypb.Any{},
-			expectError:  true,
-		},
-	}
-
-	for _, tc := range tests {
-		t.Run(tc.name, func(t *testing.T) {
-			start := time.Now()
-			buf := tu.NewMockBuffer(t)
-			params := data.CollectorParams{
-				DataType:      data.CaptureTypeTabular,
-				ComponentName: componentName,
-				Interval:      captureInterval,
-				Logger:        logging.NewTestLogger(t),
-				Clock:         clock.New(),
-				Target:        buf,
-				MethodParams:  tc.methodParams,
-			}
-
-			arm := newArm()
-			col, err := tc.collector(arm, params)
-			test.That(t, err, test.ShouldBeNil)
-
-			defer col.Close()
-			col.Collect()
-
-			ctx, cancel := context.WithTimeout(context.Background(), time.Second)
-			defer cancel()
-			if tc.expectError {
-				test.That(t, len(buf.Writes), test.ShouldEqual, 0)
-			} else {
-				tu.CheckMockBufferWrites(t, ctx, start, buf.Writes, []*datasyncpb.SensorData{{
-					Metadata: &datasyncpb.SensorMetadata{},
-					Data: &datasyncpb.SensorData_Struct{
-						Struct: tu.ToStructPBStruct(t, map[string]any{
-							"docommand_output": doCommandMap,
-						}),
-					},
-				}})
-			}
-			buf.Close()
-		})
-	}
+	tu.TestDoCommandCollector(t, tu.DoCommandTestConfig{
+		ComponentName:   componentName,
+		CaptureInterval: captureInterval,
+		DoCommandMap:    doCommandMap,
+		Collector:       arm.NewDoCommandCollector,
+		ResourceFactory: func() interface{} { return newArm() },
+	})
 }
 
 func newArm() arm.Arm {
